Create READ.me
** First Front End Mentor Project **
If you are taking the time to read this then welcome! I have been learning frontend web development slowly over the last year or so. I decided to take on some challenges through frontendmentor.io and look forward to adding more in the future.
Any/all feedback is always welcome.

-Joe